+++ resources/database.rb
@@ -26,6 +26,7 @@ property :user,     String, default: 'postgres'
 property :database, String, name_property: true
 property :host,     String
 property :port,     Integer, default: 5432
+property :psqlrc,   [true, false], default: true
 
 action :create do
   createdb = 'createdb'