+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file is used to list changes made in the last 3 major versions of the postg
 
 ## Unreleased
 
+## Unreleased
+
+- Makes psqlrc optional when invoking `psql_command_string`
+
 ## 8.1.1 - *2021-01-12*
 
 - Fix attribute updates for users with dashes